backend/requirements.txt
requests
beautifulsoup4
langchain
langchain-openai
openai
python-dotenv
fastapi
uvicorn
pydantic
tiktoken
sentence-transformers==2.2.2